+++ podcasts/PlayerContainerViewController+Update.swift
@@ -27,13 +27,11 @@ extension PlayerContainerViewController {
         let shouldShowNotes = (playingEpisode is Episode)
         let shouldShowChapters = PlaybackManager.shared.chapterCount() > 0
         let shouldShowBookmarks = true
-        let shouldShowTranscripts = FeatureFlag.transcripts.enabled && PlaybackManager.shared.transcriptsAvailable
 
         // check to see if the visible views are already configured correctly
         if shouldShowNotes == showingNotes,
             shouldShowChapters == showingChapters,
-            shouldShowBookmarks == showingBookmarks,
-            shouldShowTranscripts == showingTranscripts {
+            shouldShowBookmarks == showingBookmarks {
             return
         }
 
@@ -50,10 +48,6 @@ extension PlayerContainerViewController {
         bookmarksItem.removeFromParent()
         bookmarksItem.view.removeFromSuperview()
         showingBookmarks = false
-
-        transcriptsItem.removeFromParent()
-        transcriptsItem.view.removeFromSuperview()
-        showingTranscripts = false
 
         tabsView.tabs = [.nowPlaying]
 
@@ -78,13 +72,6 @@ extension PlayerContainerViewController {
             tabsView.tabs += [.bookmarks]
 
             addTab(bookmarksItem, previousTab: &previousTab)
-        }
-
-        if shouldShowTranscripts {
-            showingTranscripts = true
-            tabsView.tabs += [.transcript]
-
-            addTab(transcriptsItem, previousTab: &previousTab)
         }
     }