+++ podcasts/Strings+Generated.swift
@@ -382,6 +382,10 @@ internal enum L10n {
   internal static var carplayPlaybackSpeed: String { return L10n.tr("Localizable", "carplay_playback_speed") }
   /// Up Next Queue
   internal static var carplayUpNextQueue: String { return L10n.tr("Localizable", "carplay_up_next_queue") }
+  /// Thanks for being with us since the beginning! If you enjoy using our app, we’d love to hear your feedback.
+  internal static var championDescription: String { return L10n.tr("Localizable", "champion_description") }
+  /// You’re a true champion of Pocket Casts!
+  internal static var championTitle: String { return L10n.tr("Localizable", "champion_title") }
   /// Change App Icon
   internal static var changeAppIcon: String { return L10n.tr("Localizable", "change_app_icon") }
   /// Change the title that identifies this bookmark
@@ -1376,6 +1380,22 @@ internal enum L10n {
   internal static var keycommandOpenPlayer: String { return L10n.tr("Localizable", "keycommand_open_player") }
   /// Play/Pause
   internal static var keycommandPlayPause: String { return L10n.tr("Localizable", "keycommand_play_pause") }
+  /// Soon
+  internal static var kidsProfileBannerBadge: String { return L10n.tr("Localizable", "kids_profile_banner_badge") }
+  /// Request Early Access 
+  internal static var kidsProfileBannerRequestButton: String { return L10n.tr("Localizable", "kids_profile_banner_request_button") }
+  /// A new profile with only content for kids to keep you subscriptions tidy.
+  internal static var kidsProfileBannerText: String { return L10n.tr("Localizable", "kids_profile_banner_text") }
+  /// Kids Profile
+  internal static var kidsProfileBannerTitle: String { return L10n.tr("Localizable", "kids_profile_banner_title") }
+  /// No, thank you
+  internal static var kidsProfileThankyouButtonClose: String { return L10n.tr("Localizable", "kids_profile_thankyou_button_close") }
+  /// Send Feedback
+  internal static var kidsProfileThankyouButtonSend: String { return L10n.tr("Localizable", "kids_profile_thankyou_button_send") }
+  /// You’ll get a notification via the app when the new Kids profile is ready! Want to let us know what you’d like to see?
+  internal static var kidsProfileThankyouText: String { return L10n.tr("Localizable", "kids_profile_thankyou_text") }
+  /// Thank you for your interest
+  internal static var kidsProfileThankyouTitle: String { return L10n.tr("Localizable", "kids_profile_thankyou_title") }
   /// Learn More
   internal static var learnMore: String { return L10n.tr("Localizable", "learn_more") }
   /// Listening History
@@ -1749,6 +1769,10 @@ internal enum L10n {
   internal static var plusButtonTitleUnlockAll: String { return L10n.tr("Localizable", "plus_button_title_unlock_all") }
   /// Can be canceled at any time
   internal static var plusCancelTerms: String { return L10n.tr("Localizable", "plus_cancel_terms") }
+  /// Pocket Casts Champion
+  internal static var plusChampion: String { return L10n.tr("Localizable", "plus_champion") }
+  /// Thanks for being with Pocket Casts from the start. You're a real champion!
+  internal static var plusChampionMessage: String { return L10n.tr("Localizable", "plus_champion_message") }
   /// %1$@ GB Cloud Storage
   internal static func plusCloudStorageLimitFormat(_ p1: Any) -> String {
     return L10n.tr("Localizable", "plus_cloud_storage_limit_format", String(describing: p1))
@@ -1775,8 +1799,6 @@ internal enum L10n {
   internal static func plusFreeMembershipFormat(_ p1: Any) -> String {
     return L10n.tr("Localizable", "plus_free_membership_format", String(describing: p1))
   }
-  /// Lifetime Member
-  internal static var plusLifetimeMembership: String { return L10n.tr("Localizable", "plus_lifetime_membership") }
   /// Desktop & web apps
   internal static var plusMarketingDesktopAppsTitle: String { return L10n.tr("Localizable", "plus_marketing_desktop_apps_title") }
   /// Folders & Bookmarks
@@ -2168,6 +2190,8 @@ internal enum L10n {
   }
   /// Rate
   internal static var rate: String { return L10n.tr("Localizable", "rate") }
+  /// Rate Pocket Casts
+  internal static var ratePocketCasts: String { return L10n.tr("Localizable", "rate_pocket_casts") }
   /// Ops! There was an error.
   internal static var ratingError: String { return L10n.tr("Localizable", "rating_error") }
   /// Only listeners of this podcast can give it a rating. Have a listen to a few episodes and then come back to give your rating. We look forward to hearing what you think!
@@ -2422,6 +2446,8 @@ internal enum L10n {
   internal static var settingsBookmarkConfirmationSound: String { return L10n.tr("Localizable", "settings_bookmark_confirmation_sound") }
   /// Play a confirmation sound after creating a bookmark with your headphones.
   internal static var settingsBookmarkSoundFooter: String { return L10n.tr("Localizable", "settings_bookmark_sound_footer") }
+  /// Change Avatar
+  internal static var settingsChangeAvatar: String { return L10n.tr("Localizable", "settings_change_avatar") }
   /// Collect information
   internal static var settingsCollectInformation: String { return L10n.tr("Localizable", "settings_collect_information") }
   /// Allowing us to collect analytics helps us build a better app. We understand if you would prefer not to share this information.